+++ test/ppx/simple.t/run.t
@@ -4,14 +4,14 @@
     ReactIntl.{
       id: "24ac19a52c2b0192688fa8294d28dbae",
       defaultMessage: "i am message",
-      width: None,
+      maxLength: None,
     };
   let descriptedMessage: ReactIntl.message =
     [@warning "-45"]
     ReactIntl.{
       id: "6142e0d7a3dbfa8e6cbb4cfd716515fa",
       defaultMessage: "i am message",
-      width: None,
+      maxLength: None,
     };
   let message2: string =
     ReactIntlPpxAdaptor.Message.to_s(
@@ -19,7 +19,7 @@
       ReactIntl.{
         id: "900317ffb9c870f401a04bef9fd4b5b3",
         defaultMessage: "blabla",
-        width: None,
+        maxLength: None,
       },
     );
   let descriptedmessage2: string =
@@ -28,7 +28,7 @@
       ReactIntl.{
         id: "6142e0d7a3dbfa8e6cbb4cfd716515fa",
         defaultMessage: "i am message",
-        width: None,
+        maxLength: None,
       },
     );
   let undescriptedmessage: string =
@@ -37,7 +37,7 @@
       ReactIntl.{
         id: "24ac19a52c2b0192688fa8294d28dbae",
         defaultMessage: "i am message",
-        width: None,
+        maxLength: None,
       },
     );
   let element3: React.element =
@@ -47,7 +47,7 @@
         ReactIntl.{
           id: "900317ffb9c870f401a04bef9fd4b5b3",
           defaultMessage: "blabla",
-          width: None,
+          maxLength: None,
         },
       ),
     );
@@ -58,7 +58,7 @@
         ReactIntl.{
           id: "6142e0d7a3dbfa8e6cbb4cfd716515fa",
           defaultMessage: "i am message",
-          width: None,
+          maxLength: None,
         },
       ),
     );
@@ -70,7 +70,7 @@
         ReactIntl.{
           id: "4271c9d35b2eac4fca6faf3e2eeb6019",
           defaultMessage: "I am string with {variable}",
-          width: None,
+          maxLength: None,
         },
         values,
       ): string
@@ -83,7 +83,7 @@
         ReactIntl.{
           id: "a2926a901acebbdc606104ceae81dc82",
           defaultMessage: "{itemsCount, plural, zero {item} one {item} few {items} many {items} other {items}}",
-          width: None,
+          maxLength: None,
         },
         values,
       ): string
@@ -96,7 +96,7 @@
         ReactIntl.{
           id: "ac400e3c977990cd86a6981ad7eef8cd",
           defaultMessage: "I am react element with {variable}",
-          width: None,
+          maxLength: None,
         },
         values,
       );
@@ -108,7 +108,7 @@
         ReactIntl.{
           id: "a2926a901acebbdc606104ceae81dc82",
           defaultMessage: "{itemsCount, plural, zero {item} one {item} few {items} many {items} other {items}}",
-          width: None,
+          maxLength: None,
         },
         values,
       );
@@ -120,7 +120,7 @@
         ReactIntl.{
           id: "c1d9f720d6a89b19f574a7de2bf55f62",
           defaultMessage: "Some text with <a>link text</a>",
-          width: None,
+          maxLength: None,
         },
         values,
       );
@@ -142,7 +142,7 @@
           ReactIntl.{
             id: "f64fa55a351a8fe989d4fe05f15ec260",
             defaultMessage: {js|{powerUsersCountString} {powerUsersCount, plural, zero {Power users} one {Power user} few {Power users} other {Power users}}|js},
-            width: None,
+            maxLength: None,
           },
           values,
         )
@@ -181,14 +181,14 @@
       <a href="https://ocaml.com">link text</a>.
     </p>
   |js},
-        width: None,
+        maxLength: None,
       },
       values,
     );
-  let messageWithWidth: ReactIntl.message =
+  let messageWithMaxLength: ReactIntl.message =
     [@warning "-45"]
     ReactIntl.{
-      id: "4907e793cbf4bfa0bcbfe07c8d2935da",
-      defaultMessage: "i am message with width",
-      width: Some(80),
+      id: "f8dac259171ff3728f7a3317e00e265c",
+      defaultMessage: "i am message with maxLength",
+      maxLength: Some(80),
     };