+++ HISTORY.md
@@ -1,7 +1,7 @@
 # History
 
 ## 0.4.0
-- Add width field to ppx
+- Add maxLength field to ppx
 
 ## 0.3.0
 - Removed intl_draft.* extension nodes